requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
PyMuPDF==1.25.3
chromadb==0.6.3
google-generativeai==0.8.4
pydantic==2.10.4
sentence-transformers==5.4.1
requests==2.32.3
beautifulsoup4==4.13.3
